+++ test/test_codecarbon/test_codecarbon_core_functions_mysql.py
@@ -5,6 +5,7 @@ import pandas
 import pytest
 
 from py_experimenter.database_connector import DatabaseConnector
+from py_experimenter.database_connector_mysql import DatabaseConnectorMYSQL
 from py_experimenter.experimenter import PyExperimenter
 
 
@@ -15,17 +16,18 @@ def experimenter_mysql():
         os.mkdir("config")
 
     configuration_path = os.path.join("test", "test_codecarbon", "configs", "test_config_mysql.yml")
+    with patch.object(DatabaseConnectorMYSQL, "_test_connection", return_value=None):
+        experimenter = PyExperimenter(experiment_configuration_file_path=configuration_path, use_ssh_tunnel=False)
 
-    experimenter = PyExperimenter(experiment_configuration_file_path=configuration_path, use_ssh_tunnel=False)
     yield experimenter
 
     experimenter.delete_table()
 
 
 def test_delete_table_mysql(experimenter_mysql):
-    with patch.object(DatabaseConnector, "connect", return_value=None), patch.object(DatabaseConnector, "cursor", return_value=None), patch.object(
-        DatabaseConnector, "commit", return_value=None
-    ):
+    with patch.object(DatabaseConnectorMYSQL, "connect", return_value=None), patch.object(
+        DatabaseConnectorMYSQL, "close_connection", return_value=None
+    ), patch.object(DatabaseConnector, "cursor", return_value=None), patch.object(DatabaseConnector, "commit", return_value=None):
         with patch.object(DatabaseConnector, "execute", return_value=None) as mock_execute:
             experimenter_mysql.delete_table()
 
@@ -38,9 +40,9 @@ def test_delete_table_mysql(experimenter_mysql):
 
 
 def test_get_table_mysql(experimenter_mysql):
-    with patch.object(DatabaseConnector, "connect", return_value=None), patch.object(
+    with patch.object(DatabaseConnectorMYSQL, "connect", return_value=None),  patch.object(
         pandas, "read_sql", return_value=pandas.DataFrame()
-    ), patch.object(DatabaseConnector, "close_connection", return_value=None) as mock_close:
+    ), patch.object(DatabaseConnectorMYSQL, "close_connection", return_value=None) as mock_close:
         df = experimenter_mysql.get_codecarbon_table()
 
         assert df.empty is True