+++ CHANGELOG.rst
@@ -11,11 +11,10 @@ Feature
 
 - Added documentation about how to execute PyExperimenter on distributed machines.
 - Improved the usage and documentation of ssh tunnel to be more flexible and user friendly.
+- Add add_experiment_and_execute method to PyExperimenter to add and execute an experiment in one step.
 
 Fix
 ---
-
-
 
 
 v1.4.1 (11.03.2024)